// apps/pwa/src/components/pages/nestedRoutes/signs/TestKitModals/__hooks__/useTestKitModalsInfo/types.d.ts
import { BabyCheckTypeEnums, KitTestModuleTypeEnums } from '@repo/core/providers/WidgetActionsProvider';

export type TestKistModalListTypes = {
  type: 'lists';
  gregorianDate: string;
  moduleType: KitTestModuleTypeEnums;
};

export type TestKistModalBabyCheckTypes = {
  type: 'babyCheckSelectors';
  gregorianDate: string;
  babyCheckType: BabyCheckTypeEnums;
  moduleType: KitTestModuleTypeEnums;
};

export type ModalInfoTypes = null | TestKistModalListTypes | TestKistModalBabyCheckTypes;